Updated NGINX and VOD module versions in Dockerfile.
USER: I ran this command:

docker build .

And got this output:

------
 > [build 3/8] RUN nginx nginx-vod-module-1.33  && curl -sL https://github.com/kaltura/nginx-vod-module/archive/refs/tags/1.33.tar.gz | tar -C /nginx-vod-module-1.33 --strip 1 -xz:
0.254 /bin/sh: nginx: not found
------
Dockerfile:13
--------------------
  12 |
  13 | >>> RUN nginx nginx-vod-module-${VOD_MODULE_VERSION} \
  14 | >>>  && curl -sL https://github.com/kaltura/nginx-vod-module/archive/refs/tags/${VOD_MODULE_VERSION}.tar.gz | tar -C /nginx-vod-module-${VOD_MODULE_VERSION} --strip 1 -xz
  15 |
--------------------
ERROR: failed to solve: process "/bin/sh -c nginx nginx-vod-module-${VOD_MODULE_VERSION}  && curl -sL https://github.com/kaltura/nginx-vod-module/archive/refs/tags/${VOD_MODULE_VERSION}.tar.gz | tar -C /nginx-vod-module-${VOD_MODULE_VERSION} --strip 1 -xz" did not complete successfully: exit code: 127

ASSISTANT: To address the error in your Docker build, we need to make changes to the `Dockerfile`. Specifically, it seems that the `nginx` command is not found, which indicates that `nginx` is not installed or not in the PATH at the time of execution.

Please add the `Dockerfile` to the chat so I can propose the necessary changes.

USER: I added these files to the chat: Dockerfile.

If you need to propose edits to other existing files not already added to the chat, you *MUST* tell the me their full path names and ask me to *add the files to the chat*. End your reply and wait for my approval. You can keep asking if you then decide you need to edit more files.

ASSISTANT: To fix the error in the Docker build, we need to correct the command that is causing the issue. The `nginx` command is not found because it is not installed at that point in the build process. We should be downloading and extracting the `nginx-vod-module` without trying to run `nginx`.

Here are the necessary changes:

Dockerfile
```dockerfile
<<<<<<< SEARCH
RUN nginx nginx-vod-module-${VOD_MODULE_VERSION} \
=======
RUN mkdir nginx-vod-module-${VOD_MODULE_VERSION} \
>>>>>>> REPLACE
```

Please apply these changes and try building the Docker image again.